requirements.txt
numpy==1.18.2
opencv-python==4.2.0.32
Pillow==7.1.1
sacred==0.8.1
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.0
tensorboard==1.14.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.35.0
Wand==0.5.8
torchxrayvision
